% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{convert_date_and_time}
\alias{convert_date_and_time}
\title{Convert a date and time to \code{POSIXct}}
\usage{
convert_date_and_time(date, time)
}
\arguments{
\item{date}{A character vector containing dates in either 'yyyy-mm-dd' or
'mm/dd/yyyy' (note: US convension, with the month first) format.}

\item{time}{A character vector containing times in 24h format 'HH:MM' (note:
no seconds).}
}
\value{
A \link{POSIXct} object.
}
\description{
Convert a date and time to \code{POSIXct}
}
\examples{
convert_date_and_time(c("2022-01-21", "01/21/2022"), c("21:14", "9:14"))
}
