% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{convert_dtc}
\alias{convert_dtc}
\title{Convert ISO8601 date/time reprensentation to \code{POSIXct}}
\usage{
convert_dtc(x, impute = FALSE)
}
\arguments{
\item{x}{A character vector containing dates/times in their ISO8601 representation.}

\item{impute}{Whether to impute missing time information.}
}
\value{
A \link{POSIXct} object.
}
\description{
Convert ISO8601 date/time reprensentation to \code{POSIXct}
}
