% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{convert_secs}
\alias{convert_secs}
\title{Convert time in seconds since midnight to hours and minutes}
\usage{
convert_secs(x)
}
\arguments{
\item{x}{A vector containing time in seconds since midnight.}
}
\value{
A character vector containing times in 24h format 'HH:MM' (note: no seconds).
}
\description{
Convert time in seconds since midnight to hours and minutes
}
\examples{
convert_secs(1024)
}
