% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{excludeNA}
\alias{excludeNA}
\title{Shorthand Shorthand for \code{!is.na(x) & x}.}
\usage{
excludeNA(x)
}
\arguments{
\item{x}{A vector of type \code{logical}.}
}
\value{
A vector of type \code{logical}.
}
\description{
Shorthand Shorthand for \code{!is.na(x) & x}.
}
