% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{format_dtc}
\alias{format_dtc}
\title{Format a date/time object as its ISO8601 reprensentation}
\usage{
format_dtc(x)
}
\arguments{
\item{x}{A date/time object (i.e., an object which can be converted to \link{POSIXlt}).}
}
\value{
x A character vector containing the ISO8601 representation of \code{x}.
}
\description{
Format a date/time object as its ISO8601 reprensentation
}
\seealso{
\link{strftime}
}
