% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{generate_database_summary}
\alias{generate_database_summary}
\title{Generate a database summary}
\usage{
generate_database_summary(
  path.sdtm = getOption("path.sdtm", NULL),
  path.adam = getOption("path.adam", NULL),
  outfile = paste0("cdisc_database_summary_", format(Sys.time(), "\%Y\%m\%d\%H\%M\%S"),
    ".html")
)
}
\arguments{
\item{path.sdtm}{The path to the STDM files.}

\item{path.adam}{The path to the ADaM files.}

\item{outfile}{The name of the output file to generate.}
}
\value{
Called for its side effects (returns \code{NULL}).
}
\description{
Using R markdown, generate and HTML file containing a summary of the
contents of a clinical study database (STDM and ADaM), similar to a data
specification.  This is useful to search for specific information in the
absense of a true data specification.
}
\examples{
\donttest{
options(
    path.sdtm = system.file("cdiscpilot01\sdtm", package=CDISC),
    path.adam = system.file("cdiscpilot01\adam", package=CDISC))

temp <- tempfile("demo_cdisc_database_summary", fileext=".html")
generate_database_summary(outfile=temp)
browseURL(temp)
}
}
