% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{includeNA}
\alias{includeNA}
\title{Shorthand Shorthand for \code{is.na(x) | x}.}
\usage{
includeNA(x)
}
\arguments{
\item{x}{A vector of type \code{logical}.}
}
\value{
A vector of type \code{logical}.
}
\description{
Shorthand Shorthand for \code{is.na(x) | x}.
}
