% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{is0}
\alias{is0}
\title{Check for the value "0" (string or numeric), ignoring \code{NA}}
\usage{
is0(x)
}
\arguments{
\item{x}{A vector of type \code{numeric}, \code{character} or \code{factor}.}
}
\value{
A vector of type \code{logical}.
}
\description{
Check for the value "0" (string or numeric), ignoring \code{NA}
}
