% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{isND}
\alias{isND}
\title{Check for the value "NOT DONE", ignoring \code{NA}}
\usage{
isND(x)
}
\arguments{
\item{x}{A vector of type \code{character} or \code{factor}.}
}
\value{
A vector of type \code{logical}.
}
\description{
Check for the value "NOT DONE", ignoring \code{NA}
}
