% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{isY}
\alias{isY}
\title{Check for the value "Y", ignoring \code{NA}}
\usage{
isY(x)
}
\arguments{
\item{x}{A vector of type \code{character} or \code{factor}.}
}
\value{
A vector of type \code{logical}.
}
\description{
Check for the value "Y", ignoring \code{NA}
}
