% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{merge_supp}
\alias{merge_supp}
\title{Merge SDTM "supp" data}
\usage{
merge_supp(base, supp, .lc = TRUE)
}
\arguments{
\item{base}{A \code{data.frame} containing the "base" data (e.g. vc, lb, pc, ...).}

\item{supp}{A \code{data.frame} containing the "supp" data (e.g. suppvc, supplb, supppc, ...).}

\item{.lc}{If \code{TRUE}, then convert all new column names added to \code{base} to lower case.}
}
\value{
A \code{data.frame}.
}
\description{
Merge SDTM "supp" data
}
