% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{partial_date}
\alias{partial_date}
\title{Handle partial dates}
\usage{
partial_date(x, impute = c("missing", "first", "mid"))
}
\arguments{
\item{x}{A character vector containing a representation of a vector of dates
in \code{yyyy-mm-dd} format, some of which may be incomplete.}

\item{impute}{Whether and how to impute the missing information (see Details).}
}
\value{
A \link{Date} object.
}
\description{
In CDISC datasets, dates can sometimes be incomplete due to missing
information.
}
