% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{read_adam}
\alias{read_adam}
\title{Read ADaM data}
\usage{
read_adam(
  domain,
  path = getOption("path.adam", "."),
  extension = "(sas7bdat|xpt)",
  verbose = TRUE
)
}
\arguments{
\item{domain}{An ADaM "domain" (usually 4 letters starting with "ad", such as "adsl" or "adae").}

\item{path}{A path in which to search for ADaM files.}

\item{extension}{File extension(s) to search for (can be a regular expression).}

\item{verbose}{If \code{TRUE} (the default), output informative messages.}
}
\value{
A \code{data.frame}.
}
\description{
Read ADaM data
}
