% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{read_data}
\alias{read_data}
\title{Helper function to read data from files}
\usage{
read_data(file, path = NULL, ..., .fun = NULL, .lc = TRUE, .dt = TRUE)
}
\arguments{
\item{path}{Optionally, a path in which to look for \code{file}.}

\item{...}{Further arguments passed to the function that does the actual reading.}

\item{.fun}{A function that reads a file passed as its first argument. The
default is to use an appropriate function depending on the file extension.}

\item{.lc}{If \code{TRUE}, then convert all names to lower case in the returned object.}

\item{.dt}{If \code{TRUE}, then then returned a \code{data.table}.}

\item{x}{A file name.}
}
\value{
If \code{.dt} is \code{TRUE}, then a \code{data.table} is returned. Otherwise, a
\code{data.frame} is returned unless \code{.fun} is specified, in which case that
function determines the return type.
}
\description{
Helper function to read data from files
}
