% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{read_data_excel}
\alias{read_data_excel}
\title{Helper function to read data from Excel files}
\usage{
read_data_excel(file, path = NULL, ...)
}
\arguments{
\item{path}{Optionally, a path in which to look for \code{file}.}

\item{...}{Further arguments passed to the function that does the actual reading.}

\item{x}{A file name.}
}
\value{
A \code{data.frame}.
}
\description{
Helper function to read data from Excel files
}
