% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc.R
\name{safe_rbind}
\alias{safe_rbind}
\title{Safely bind rows}
\usage{
safe_rbind(..., .prefer = c("factor", "character", "error"))
}
\arguments{
\item{...}{Objects (typically \code{data.frame}s) that contain rows to bind.}

\item{.prefer}{What to do when column types don't match (such as when the
same column is numeric in one \code{data.frame} and textual in another).}
}
\value{
A \code{data.frame}.
}
\description{
Safely bind rows
}
\seealso{
\link{rbind}
}
